backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
python-dotenv>=1.1,<2.0
dj-database-url>=2.3,<3.0
psycopg[binary]>=3.2,<4.0
gunicorn>=23.0,<24.0
